Реализовал систему подсчета оценки отзывов игр

+++ LingvoGameOsWeb/Areas/Admin/Controllers/ReviewsController.cs
@@ -1,5 +1,6 @@
 ﻿using LingvoGameOs.Db;
 using LingvoGameOs.Db.Models;
+using LingvoGameOs.Helpers;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
@@ -13,10 +14,13 @@ namespace LingvoGameOs.Areas.Admin.Controllers
     {
         readonly IReviewsRepository _reviewsRepository;
         readonly UserManager<User> _userManager;
-        public ReviewsController(IReviewsRepository reviewsRepository, UserManager<User> userManager)
+        readonly RatingService _ratingService;
+
+        public ReviewsController(IReviewsRepository reviewsRepository, UserManager<User> userManager, RatingService ratingService)
         {
             _reviewsRepository = reviewsRepository;
             _userManager = userManager;
+            _ratingService = ratingService;
         }
 
         public async Task<IActionResult> Index()
@@ -26,8 +30,8 @@ namespace LingvoGameOs.Areas.Admin.Controllers
                 return NotFound();
 
             var reviews = await _reviewsRepository.GetAllAsync();
-            int approvedReviewsCount =reviews.Count(r => r.IsApproved == true);
-            
+            int approvedReviewsCount = reviews.Count(r => r.IsApproved == true);
+
             ViewBag.ApprovedReviewsCount = approvedReviewsCount;
             ViewBag.NotApprovedReviewsCount = reviews.Count - approvedReviewsCount;
             ViewBag.UsersCount = _userManager.Users.Count();
@@ -41,7 +45,14 @@ namespace LingvoGameOs.Areas.Admin.Controllers
         {
             try
             {
+                var existingReview = await _reviewsRepository.TryGetByIdAsync(reviewId);
+                if (existingReview == null)
+                    return BadRequest(new { success = false, message = "Отзыв с таким id не найден!" });
+
                 await _reviewsRepository.PublishAsync(reviewId);
+
+                await _ratingService.UpdateGameRatingAsync(existingReview.GameId, existingReview.Rating);
+
                 return Ok(new
                 {
                     success = true
@@ -62,6 +73,11 @@ namespace LingvoGameOs.Areas.Admin.Controllers
         {
             try
             {
+                var existingReview = await _reviewsRepository.TryGetByIdAsync(reviewId);
+                if (existingReview == null)
+                    return BadRequest(new { success = false, message = "Отзыв с таким id не найден!" });
+
+                await _ratingService.CalculateRatingAfterDeleteAsync(existingReview.GameId, existingReview.Rating);
                 await _reviewsRepository.DeleteAsync(reviewId);
                 return Ok(new
                 {
